layout/base/tests/test_synthmousemove.html

<!doctype html>
<title>test synth mouse moves go to the right place with fission</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script src="/tests/SimpleTest/EventUtils.js"></script>

<script>
SimpleTest.waitForExplicitFinish();

function runTest() {
  window.open("file_synthmousemove.html", "_blank");
}

function reportSuccess() {
  ok(true, "reportSuccess");
  SimpleTest.finish();
}

function prepareTest() {
  window.requestAnimationFrame(function() {
    SpecialPowers.pushPrefEnv({"set":[["general.smoothScroll", false]]}, runTest);
  });
}

SimpleTest.waitForFocus(prepareTest);
</script>
